Move helper functions from views.py to util.py
Route files should be a clean as poss
