+++ orlo/error_handlers.py
@@ -1,0 +1,28 @@
+from __future__ import print_function
+from flask import jsonify, request
+from orlo import app
+from orlo.exceptions import InvalidUsage
+
+__author__ = 'alforbes'
+
+
+@app.errorhandler(404)
+def page_not_found(error):
+    d = error.to_dict()
+    d['url'] = request.url
+    return jsonify(d), 404
+
+
+@app.errorhandler(InvalidUsage)
+def handle_invalid_usage(error):
+    response = jsonify(error.to_dict())
+    response.status_code = error.status_code
+    return response
+
+
+@app.errorhandler(400)
+def handle_400(error):
+    response = jsonify(error.to_dict())
+    response.status_code = error.status_code
+    return response
+
